Update to v083r06 release.
byuu says:

All cores: Video classes have internal->{RGB30,24,16,15} palette
generation support
All cores: video output is now RGB24, all filters except HQ2x were
updated to reflect this (HQ2x will be very hard)
NES: MMC5 CHR mapping fixes (Bandit Kings, RTK2, Uchuu Keibitai SDF)
[Cydrak]
NES: MMC5 vertical split screen support (Uchuu Keibitai SDF) [Cydrak]
Game Boy + Game Boy Color: fixed a potential freezing bug when loading
save states (re-create cothreads on state load; was implied when using
SGB mode.)
Game Boy Color: fixed freezing bug with Zelda: LA opening (SVBK is
readable.)
Game Boy Color: more accurate colors (better than GiiBii, probably worse
than KiGB)
SNES: luminance of zero is no longer pure black, as on real hardware.
This is possible thanks to using RGB888 output now.

The current major problems I'd like to solve:
- Zelda: Link's Awakening music when Link first wakes up in the house is
  atrociously bad
- Shin Megami Tensei: Devil Children - White Book (Shiro no Sho) plays
  music at 50% speed; yet Black Book (Kuro no Sho) does not ... one of
  my favorite games, so it'd be great to fix it

+++ bsnes/ui/interface/interface.hpp
@@ -6,12 +6,12 @@
 
 struct Filter : public library {
   function<void (unsigned&, unsigned&)> dl_size;
-  function<void (uint16_t*, unsigned, const uint16_t*, unsigned, unsigned, unsigned)> dl_render;
-  void render(const uint16_t*, unsigned, unsigned, unsigned);
+  function<void (uint32_t*, unsigned, const uint32_t*, unsigned, unsigned, unsigned)> dl_render;
+  void render(const uint32_t*, unsigned, unsigned, unsigned);
   Filter();
   ~Filter();
 
-  uint16_t *data;
+  uint32_t *data;
   unsigned pitch;
   unsigned width;
   unsigned height;
@@ -47,7 +47,7 @@ struct Interface : property<Interface> {
   Interface();
 
   bool loadFile(const string &filename, uint8_t *&data, unsigned &size);
-  void videoRefresh(const uint16_t *input, unsigned inputPitch, unsigned width, unsigned height);
+  void videoRefresh(const uint32_t *input, unsigned inputPitch, unsigned width, unsigned height);
 
   string baseName;  // = "/path/to/cartridge" (no extension)
   lstring slotName;
